Design Complete Draft 1

+++ commerce/auctions/views.py
@@ -1,4 +1,5 @@
 from cProfile import label
+from doctest import BLANKLINE_MARKER
 import re
 from unicodedata import category
 from django.contrib.auth import authenticate, login, logout
@@ -18,7 +19,7 @@ class addListingForm(forms.Form):
     title = forms.CharField(label="Title")
     starting_bid = forms.IntegerField(label="Starting Bid")
     category = forms.CharField(label="Category")
-    img_link = forms.URLField(label="Image Link")
+    img_link = forms.URLField(label="Image Link", required=False)
     description = forms.CharField(label="Description", widget=forms.Textarea(attrs={'rows': 4, 'cols':100}))
 
 
@@ -117,36 +118,38 @@ def listing_view(request, listing_id):
     
     if request.method == "POST":
 
-        if ("bid" in request.POST):
+        if (curr_user.is_authenticated):
 
-            bid_amount = int(request.POST["bid"])
+            if ("bid" in request.POST):
 
-            if (bid_amount > highest_bid(list_item)): # Get highest bid again for server-sdie check in case another bid was successful before the client-side is updated.
-                            
-                new_bid = Bid(bid=bid_amount, bidder=curr_user, listing=list_item)
-                new_bid.save()
+                bid_amount = int(request.POST["bid"])
 
-
-        elif ("addtowatchlist" in request.POST):
-
-            list_item.watchlist.add(curr_user)
-
-        elif ("removefromwatchlist" in request.POST):
-
-            list_item.watchlist.through.objects.filter(user_id = curr_user.id).delete() 
-
-        elif ("closeauction" in request.POST):
-        
-            list_item.active = False
-            list_item.save()
-        
-        elif ("comment" in request.POST):
-
-            new_comment = Comment(creator=curr_user,listing=list_item,comment=request.POST["comment"])
-            new_comment.save()
+                if (bid_amount > highest_bid(list_item)): # Get highest bid again for server-sdie check in case another bid was successful before the client-side is updated.
+                                
+                    new_bid = Bid(bid=bid_amount, bidder=curr_user, listing=list_item)
+                    new_bid.save()
 
 
-        return HttpResponseRedirect(reverse("listing", args=[listing_id]))
+            elif ("addtowatchlist" in request.POST):
+
+                list_item.watchlist.add(curr_user)
+
+            elif ("removefromwatchlist" in request.POST):
+
+                list_item.watchlist.through.objects.filter(user_id = curr_user.id).delete() 
+
+            elif ("closeauction" in request.POST):
+            
+                list_item.active = False
+                list_item.save()
+            
+            elif ("comment" in request.POST):
+
+                new_comment = Comment(creator=curr_user,listing=list_item,comment=request.POST["comment"])
+                new_comment.save()
+
+
+            return HttpResponseRedirect(reverse("listing", args=[listing_id]))
 
 
     return render(request, "listing/index.html",{
